resource/requirements.txt
python-dotenv==0.10.1
Flask==1.1.1
requests==2.20.0
itsdangerous==2.0.1
werkzeug==2.0.3
